factor out the temporary lua files in tests
